diffusion-cifar10/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
lpips>=0.1.4
pytorch-fid>=0.3.0 